pyproject.toml
[project]
name = "academic-paper-summarizer"
version = "0.1.0"
description = "RAG-powered academic paper summarizer with Gemini and ScaleDown APIs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.31.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.4",
    "pymupdf>=1.23.0",
    "google-generativeai>=0.4.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "alembic>=1.13.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.20.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["app"]
omit = ["app/pages/*"]
